Update git helpers
Many repositories use submodules for lots of different things
we don't use. We end up cloning the world and having duplicated
repositories all over the place. Also this way it takes forever
to clone/update the repositories and for some repositories
we can't do a shallow clone (e.g. qemu).

To prevent this introduce a new helper for manualy checking out
submodules and allow to disable recursive cloning.