Toolchain handling updates
Instead of recursively cloning the riscv-gnu-toolchain, manualy
clone the needed submodules and also do a shallow checkout to
reduce the time and space it takes.

Also make sure that during bootstrap if a step fails the
process stops, this preserves the logs so that we can see
what's going on.